Added logout support + fixed CORSFilter issues + added some stuff
