--- src/main/java/homies/goats/projet_jee/CORSFilter.java
@@ -1,39 +1,0 @@
-package homies.goats.projet_jee;
-
-import jakarta.servlet.*;
-import jakarta.servlet.annotation.WebFilter;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-@WebFilter("/*")
-public class CORSFilter implements Filter {
-    @Override
-    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        HttpServletResponse servletResponse = (HttpServletResponse) response;
-        servletResponse.setHeader("Access-Control-Allow-Origin", "*");
-        servletResponse.setHeader("Access-Control-Allow-Methods", "POST, GET, PUT, OPTIONS, DELETE");
-        servletResponse.setHeader("Access-Control-Max-Age", "3600");
-        servletResponse.setHeader("Access-Control-Allow-Headers", "x-requested-with, Content-Type");
-        chain.doFilter(request, response);
-
-        HttpServletRequest httpRequest = (HttpServletRequest) request;
-        String requestURI = httpRequest.getRequestURI();
-
-        if (httpRequest.getSession().getAttribute("authenticatedUser") == null) {
-            if (!requestURI.endsWith("/signup.jsp") || !requestURI.endsWith("/login.jsp")){
-                return;
-            } else {
-                HttpServletResponse httpResponse = (HttpServletResponse) response;
-                httpResponse.sendRedirect(httpRequest.getContextPath() + "/login.jsp");
-            }
-        }
-    }
-    @Override
-    public void init(FilterConfig filterConfig) {
-    }
-    @Override
-    public void destroy() {
-    }
-}
