+++ src/main/webapp/WEB-INF/web.xml
@@ -35,12 +35,31 @@
         <url-pattern>/gestion</url-pattern>
     </servlet-mapping>
     
+    <servlet>
+        <servlet-name>LogoutServlet</servlet-name>
+        <servlet-class>homies.goats.projet_jee.servlet.LogoutServlet</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>LogoutServlet</servlet-name>
+        <url-pattern>/logout</url-pattern>
+    </servlet-mapping>
+    
     <filter>
         <filter-name>CORSFilter</filter-name>
-        <filter-class>homies.goats.projet_jee.CORSFilter</filter-class>
+        <filter-class>homies.goats.projet_jee.filters.CORSFilter</filter-class>
     </filter>
     <filter-mapping>
         <filter-name>CORSFilter</filter-name>
         <url-pattern>/*</url-pattern>
     </filter-mapping>
+
+    <filter>
+        <filter-name>AuthenticationFilter</filter-name>
+        <filter-class>homies.goats.projet_jee.filters.AuthenticationFilter</filter-class>
+    </filter>
+    <filter-mapping>
+        <filter-name>AuthenticationFilter</filter-name>
+        <url-pattern>/*</url-pattern>
+    </filter-mapping>
+
 </web-app>
