+++ src/main/webapp/gestion.jsp
@@ -16,109 +16,111 @@
 
 </head>
 <body>
-<header>
-  <div class="left">
-    <img src="resources/images/iron.png" alt="Logo du site">
-    <h1>Trisotto</h1>
-  </div>
-  <div class="right">
-    <div class="user-dropdown">
-      <img src="resources/images/lol.png" alt="Icône utilisateur">
-      <div class="dropdown-content">
-        <a href="parameters.jsp">Paramètres</a>
-        <a href="#">Déconnexion</a>
+  <header>
+    <div class="left">
+      <img src="resources/images/iron.png" alt="Logo du site">
+      <h1>Bonjour ${authenticatedUser.forename} ${authenticatedUser.lastname}</h1>
+    </div>
+    <div class="right">
+      <div class="user-dropdown">
+        <img src="resources/images/lol.png" alt="Icône utilisateur">
+        <div class="dropdown-content">
+          <a href="parameters.jsp">Paramètres</a>
+          <a href="${pageContext.request.contextPath}/logout">Déconnexion</a>
+        </div>
       </div>
     </div>
-  </div>
-</header>
-<!-- Le reste du contenu de votre page -->
-</body>
-<h1>List of Users</h1>
-<div class="researchTable">
-  <div class="inputs">
+  </header>
 
-    <div class="filter-input-container">
-      <i class="fas fa-search"></i> <!-- Icône de loupe (Font Awesome) -->
-      <input type="text" placeholder="Filtrer">
+  <h1>List of Users</h1>
+  <div class="researchTable">
+    <div class="inputs">
+
+      <div class="filter-input-container">
+        <i class="fas fa-search"></i> <!-- Icône de loupe (Font Awesome) -->
+        <input type="text" placeholder="Filtrer">
+      </div>
+
+      <div class="buttons">
+        <input type="submit" value="Tutor" class="submit buttonFilter">
+        <input type="submit" value="Apprentice" class="submit buttonFilter">
+      </div>
     </div>
 
-    <div class="buttons">
-      <input type="submit" value="Tutor" class="submit buttonFilter">
-      <input type="submit" value="Apprentice" class="submit buttonFilter">
-    </div>
-  </div>
-
-  <div class="tablo">
-    <table>
-      <thead>
-      <tr>
-        <th>Name</th>
-        <th>Forename</th>
-        <th>Email</th>
-        <th>Phone</th>
-        <th>Role</th>
-        <th></th>
-      </tr>
-      </thead>
-
-      <c:if test="${not empty USER_LIST}">
-        <c:forEach var="user" items="${USER_LIST}">
+    <c:if test="${not empty USER_LIST}">
+      <div class="tablo">
+        <table>
+          <thead>
           <tr>
-            <td>${user.lastname}</td>
-            <td>${user.forename}</td>
-            <td>${user.email}</td>
-            <td>${user.phone}</td>
-            <td>${user.userType}</td>
-            <td><button class="submit" onclick="openPopup()">Details</button></td>
+            <th>Name</th>
+            <th>Forename</th>
+            <th>Email</th>
+            <th>Phone</th>
+            <th>Role</th>
+            <th></th>
           </tr>
-        </c:forEach>
+          </thead>
+
+            <c:forEach var="user" items="${USER_LIST}">
+              <tr>
+                <td>${user.lastname}</td>
+                <td>${user.forename}</td>
+                <td>${user.email}</td>
+                <td>${user.phone}</td>
+                <td>${user.userType}</td>
+                <td><button class="submit" onclick="openPopup()">Details</button></td>
+              </tr>
+            </c:forEach>
+
+        </table>
       </c:if>
-    </table>
-  </div>
-</div>
-
-<c:if test="${empty USER_LIST}">
-  <p>No data found.</p>
-</c:if>
-
-<div id="popup" class="popup">
-  <div class="popup-content">
-    <span class="close" id="close-popup" onclick="closePopup()">&times;</span>
-    <h2 class="popup-title">Ineed Morebullets</h2>
-    <div class="input-row">
-      <div class="half-width flex">
-        <span class="bold">Academic Year</span>
-        <span>2022 - 2023</span>
-      </div>
-      <div class="half-width flex">
-        <span class="bold">Program</span>
-        <span>RS</span>
-      </div>
-    </div>
-    <div class="input-row">
-      <div class="half-width flex">
-        <span class="bold">Major</span>
-        <span>Cybersecurity</span>
-      </div>
-      <div class="half-width flex">
-        <span class="bold">Tutor</span>
-        <span>LaFrappeDeChine</span>
-      </div>
-    </div>
-    <hr />
-    <h3 class="popup-title"> US Navy </h3>
-    <div class="input-row">
-      <div class="half-width flex">
-        <span class="bold">Address</span>
-        <span>4 rue des Chasseurs</span>
-      </div>
-      <div class="half-width flex">
-        <span class="bold">Infos</span>
-        <span>A la cambrousse</span>
-      </div>
+      <c:if test="${empty USER_LIST}">
+        <p>La liste est vide. Ajoutez au moins un(e) apprenti(e)</p>
+      </c:if>
     </div>
   </div>
-</div>
+
+
+
+  <div id="popup" class="popup">
+    <div class="popup-content">
+      <span class="close" id="close-popup" onclick="closePopup()">&times;</span>
+      <h2 class="popup-title">Ineed Morebullets</h2>
+      <div class="input-row">
+        <div class="half-width flex">
+          <span class="bold">Academic Year</span>
+          <span>2022 - 2023</span>
+        </div>
+        <div class="half-width flex">
+          <span class="bold">Program</span>
+          <span>RS</span>
+        </div>
+      </div>
+      <div class="input-row">
+        <div class="half-width flex">
+          <span class="bold">Major</span>
+          <span>Cybersecurity</span>
+        </div>
+        <div class="half-width flex">
+          <span class="bold">Tutor</span>
+          <span>LaFrappeDeChine</span>
+        </div>
+      </div>
+      <hr />
+      <h3 class="popup-title"> US Navy </h3>
+      <div class="input-row">
+        <div class="half-width flex">
+          <span class="bold">Address</span>
+          <span>4 rue des Chasseurs</span>
+        </div>
+        <div class="half-width flex">
+          <span class="bold">Infos</span>
+          <span>A la cambrousse</span>
+        </div>
+      </div>
+    </div>
+  </div>
+</body>
 <script>
   function closePopup() {
     var popup = document.getElementById("popup");
